Split the public API route tests off the max-lines list
Break test/routes/api.test.ts (1,320 lines) into a test/routes/api/ folder
grouped by endpoint, and drop the file's entry from the
noExcessiveLinesPerFile ignore list.

The single "Public API" describeWithEnv wrapper, its beforeEach, and the
dozen shared request helpers (bookListing, fetchListingBySlug,
withCheckoutStub, …) move to helpers.ts behind a describePublicApi() wrapper
each shard reuses. The huge POST .../book suite is split by theme into
book-basics, book-payments and book-parents.

Bringing the file back under the duplication checker's size limit surfaced
booking setup that was repeated inside the monolith; that's now shared
fixtures (bookForToken, bookTwoChildEntries, captureCheckoutIntent,
targetAndDecoy).

Every original test carries over unchanged; the folder's 97 nodes pass, and
the linter and duplication checker are clean.
